People Page UI update
People page is now fully responsive.
Added content in people.html into d-flex and added columns as needed.
Renamed user in object_list to profile in object_list for clarity.
Fixed missing img-responsive in dashboard.html
Added shrink-to-fit to meta tag as per Bootstrap 4 recommendations
In user-cards.css added a margin of 1% to the bottom of cards, ensures cards are always on a shadowed background.
In base.css I cleaned up spacing, removed main-div styling (currently was not being used), and added a .25em margin to top and bottom on all .btn as well as  ensuring the button min-width is at least the width if its content.