Profile and thumbnail UI update
Updated profile to be fully responsive as well as adjusted image thumbnails.
Removed img-responsive from CSS as it is unneeded. Renamed to img-thumbnail in HTML.
Ensured a min-width of 175% is present in the edit photo only via style.
Removed img-responsive from people.html as it is no longer needed.
Added d-flex and columns to profile page as needed.