Initial Comment System
Currently comments are saved to the database only (only available in admin).
Added Comment model to post/models.py.
Parent is for anything the comment is attached to and is intended to allow sub-comments.
Also added user, body and a DateTimeField for created.
Added CommentForm to post/forms.py. The only field is body.
Reorganized imports in post/views.py. Nowimport FormView, HttpResponseForbidden, get_object_or_404, View, SingleObjectMixin, Comment and CommentForm.

Re-wrote PostDetailView to do the following:
On GET requests it returns PostDisplayView as_view and shows the post_detail template.
On POST requests it returns CommentFormView as_view and shows reloads the page with the comment saved to the database.
Currently post details does not show the actual comments just admin.