added next and previous tests
Added testing for next and previous actions, both tests passing.
Added page_num to post method of GetFollowers class to fix failing test.
In previous method of PaginateManager, removed multiplication on next_set as it was unneeded and causing a test to fail. Also swapped order of next set and previous set.